pyproject.toml
[project]
name = "learn-deepagents"
version = "0.1.0"
description = "Study logs of langchain deep agents, model-agnostic alternative to Claude Agents SDK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.3",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]
